+++ HISTORY.rst
@@ -24,6 +24,7 @@ History
 * BODY-38: REST Resources: notify events for the model instance lifecycle (POST, PUT, GET, DELETE) (rudaporto)
 * BODY-39: Fix workflow resource POST transition (rudaporto)
 * BODY-44: Public information about a user. (ericof)
+* BODY-47: Add sqlalchemy listener events to inject request on model creation and load. (rudaporto)
 
 
 0.1.0 (Unreleased)